+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
+## [Unreleased]
+
+### Changed
+
+### Added
+
+- `tax`: `Scenario` now has `Filter` property to set a code function
+
 ## [v0.115.1]
 
 ### Fixes